+++ src/Speckle.Sdk/Serialisation/V2/Receive/ObjectLoader.cs
@@ -32,7 +32,6 @@ public sealed class ObjectLoader(
         .ConfigureAwait(false);
       sqLiteCacheManager.SaveObjectSync(objectId, rootJson);
     }
-
     return rootJson;
   }
