+++ src/Speckle.Sdk/Transports/ProgressArgs.cs
@@ -1,0 +1,14 @@
+namespace Speckle.Sdk.Transports;
+
+public readonly record struct ProgressArgs(ProgressEvent ProgressEvent, long Count, long? Total);
+
+public enum ProgressEvent
+{
+  CacheCheck,
+  DownloadBytes,
+  UploadBytes,
+  DownloadObject,
+  UploadObject,
+  DeserializeObject,
+  SerializeObject,
+}
