+++ tests/Speckle.Sdk.Serialization.Testing/Progress.cs
@@ -7,16 +7,28 @@ public class Progress(bool write) : IProgress<ProgressArgs>
   private readonly TimeSpan DEBOUNCE = TimeSpan.FromMilliseconds(500);
   private DateTime _lastTime = DateTime.UtcNow;
 
+  private long _totalBytes;
+
   public void Report(ProgressArgs value)
   {
     if (write)
     {
+      if (value.ProgressEvent == ProgressEvent.DownloadBytes)
+      {
+        Interlocked.Add(ref _totalBytes, value.Count);
+      }
       var now = DateTime.UtcNow;
       if (now - _lastTime >= DEBOUNCE)
       {
-        Console.WriteLine(
-          value.ProgressEvent + " " + value.Count + " " + value.Total + " " + Environment.CurrentManagedThreadId
-        );
+        if (value.ProgressEvent == ProgressEvent.DownloadBytes)
+        {
+          Console.WriteLine(value.ProgressEvent + " t " + _totalBytes);
+        }
+        else
+        {
+          Console.WriteLine(value.ProgressEvent + " c " + value.Count + " t " + value.Total);
+        }
+
         _lastTime = now;
       }
     }
